rename to legacy
we'll duplicate and reuse the existing view where appropriate
