+++ subscriptions/subscriptions-impl/src/main/java/com/duckduckgo/subscriptions/impl/settings/plugins/SubsSettingsPlugins.kt
@@ -21,11 +21,12 @@ import android.view.View
 import com.duckduckgo.anvil.annotations.PriorityKey
 import com.duckduckgo.common.ui.view.listitem.SectionHeaderListItem
 import com.duckduckgo.di.scopes.ActivityScope
+import com.duckduckgo.settings.api.NewSettingsFeature
 import com.duckduckgo.settings.api.ProSettingsPlugin
 import com.duckduckgo.subscriptions.impl.R
 import com.duckduckgo.subscriptions.impl.settings.views.ItrSettingView
 import com.duckduckgo.subscriptions.impl.settings.views.PirSettingView
-import com.duckduckgo.subscriptions.impl.settings.views.ProSettingView
+import com.duckduckgo.subscriptions.impl.settings.views.LegacyProSettingView
 import com.squareup.anvil.annotations.ContributesMultibinding
 import javax.inject.Inject
 
@@ -43,7 +44,7 @@ class ProSettingsTitle @Inject constructor() : ProSettingsPlugin {
 @PriorityKey(500)
 class ProSettings @Inject constructor() : ProSettingsPlugin {
     override fun getView(context: Context): View {
-        return ProSettingView(context)
+        return LegacyProSettingView(context)
     }
 }
