+++ subscriptions/subscriptions-impl/src/main/res/layout/legacy_view_settings.xml
